+++ include/konoha3/platform_msvc.h
@@ -232,7 +232,7 @@ static void loadI18N(KonohaFactory *plat, const char *defaultCharSet)
 
 // -------------------------------------------------------------------------
 
-static unsigned long long getTimeMilliSecond(void)
+static uint64_t getTimeMilliSecond(void)
 {
 	SYSTEMTIME st;
 	GetSystemTime(&st);
